+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-315](https://github.com/itk-dev/economics/pull/315)
+  * Add tidy-feedback collector module.
 * [PR-313](https://github.com/itk-dev/economics/pull/313)
   * Changed supervisor to php 8.4 version.
 * [PR-310](https://github.com/itk-dev/economics/pull/310)